requirements.txt
apache-airflow
psycopg2-binary
requests
great-expectations
pandas
unidecode
